requirements.txt
Flask
flask-cors
flask-sqlalchemy
pyjwt
flask_login